- Fix: MapStruct and Lombok conflict on the builder method
Added -proc:none as compiler arg for solution

+++ src/main/java/com/eduardoandre/booksstoremanager/entities/Author.java
@@ -1,10 +1,7 @@
 package com.eduardoandre.booksstoremanager.entities;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Entity
 @Data
@@ -16,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @Column(nullable = false, unique = true)
     private String name;
